Added early basic shadowmapping (WIP)
Shadowmap generation utilizes now the GPU driven culling and rendercommand creation pipeline. The pipeline is versatile enough to take a view and generate a rendercommand list for each view. On higher level we can maintain these views (like storing 6 views for a pointlight shadowmap component).

Still early tho and limited. This is just an intermediate submit as well not to lose current progress.

Next:
- Remove limitation on view count (current limited to 32 views which is too few)
- Bring animations back in
- Do a lot code clean up