- Culling and render command pass (ComputePass) generates a list which contains all visible entity ids - added pass to update dynamic data and animation data (using given list of visible entities) - cleaned some shadercode a bit and removed few unncessary files - added simple filmic tonemapping and gamma adjustment for srgb.
The compute pass works fine now with an arbitrary number of views, for shadowmapping and mulltiple cameras (buffer still dont grow dynamically but thats not a problem). Given approach allows to implement further culling technics like high-z culling which directly affects all views.

However, there is still room for improvement. The generated instances lists from the compute pass are still sorted on CPU - needs to be done on GPU later.
Writes and Readbacks of compute shader could eventually be done without barriers using ring buffers or similar. This wouldnt stall the CPU and GPU.
The culling also renders instances for pointlights and views which are not even close to be visible. So shadow culling could be something.
Still need to implement proper viewgroup and viewid maintanance and entity data unloading.